Get article by Id is in progress

+++ src/main/java/com/demo/postgresqlspring/domain/article/repository/ArticleRepository.java
@@ -3,6 +3,6 @@ package com.demo.postgresqlspring.domain.article.repository;
 import com.demo.postgresqlspring.domain.article.entity.ArticleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<ArticleEntity, Long> {
+public interface ArticleRepository extends JpaRepository<ArticleEntity, Integer> {
 
 }
